+++ adapters/MyTarget/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## myTarget iOS Mediation Adapter Changelog
 
+#### Next Version
+- Added standardized adapter error codes and messages.
+
 #### Version 5.9.5.0
 - Verified compatibility with myTarget SDK version 5.9.5.
 
@@ -74,7 +77,7 @@ Build and tested With:
 Build and tested With:
 - Google Mobile Ads SDK version 7.61.0.
 - MyTarget SDK version 5.7.0
-.
+
 #### Version 5.6.3.0
 - Verified compatibility with myTarget SDK version 5.6.3.
 - Now requires Google Mobile Ads SDK version 7.61.0 or higher.